+++ CHANGELOG.md
@@ -36,6 +36,6 @@ The subsection are to be interpreted as defined [here](http://keepachangelog.com
 ## 0.1.0 - 2015-02-04
 Initial release
 
-
+[unreleased]: https://github.com/baig/brackets-zotero/compare/0.1.2...HEAD
 [0.1.2]: https://github.com/baig/brackets-zotero/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/baig/brackets-zotero/compare/0.1.0...0.1.1